Fix task-service BigInteger/INTEGER schema mismatch for work logs
TaskWorkLog entity: BigInteger → Integer for plannedHours/bookedHours
to match the INTEGER column type in V5 migration (same fix as audit-service).
Conversions: request BigInteger → Integer on save; Integer → BigInteger on
toResponse; BigInteger.valueOf(int) when building Kafka outbox events.
Also fixes TaskWorkLogControllerIT assertion: plannedHours after update
should be 4 (the immutable original), not 2 (the rejected update value).

+++ task-service/src/main/java/com/demo/task/model/TaskWorkLog.java
@@ -6,7 +6,6 @@ import lombok.*;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.SQLRestriction;
 
-import java.math.BigInteger;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -40,11 +39,11 @@ public class TaskWorkLog {
 
     /** Estimated hours planned for this work entry. */
     @Column(name = "planned_hours", nullable = false)
-    private BigInteger plannedHours;
+    private Integer plannedHours;
 
     /** Actual hours worked and booked against this entry. */
     @Column(name = "booked_hours", nullable = false)
-    private BigInteger bookedHours;
+    private Integer bookedHours;
 
     @Column(name = "created_at", nullable = false)
     private Instant createdAt;
